1. @staticmethod
