+++ authapp/templates/authapp/edit.html
@@ -83,6 +83,7 @@
                     </div>
                     <br>
 
+
                     <input class="btn-inverse form-control float-left"
                            type="submit"
                            value="Сохранить">
